web: fix export issue in CI
