+++ browser/src/shared/tracking/eventLogger.tsx
@@ -6,8 +6,7 @@ import { PlatformContext } from '../../../../shared/src/platform/context'
 import { TelemetryService } from '../../../../shared/src/telemetry/telemetryService'
 import { storage } from '../../browser/storage'
 import { isInPage } from '../../context'
-import { logUserEvent } from '../backend/userEvents'
-import { UserEvent } from '../../../../web/src/user/settings/backend'
+import { logUserEvent, UserEvent } from '../backend/userEvents'
 import { observeSourcegraphURL } from '../util/context'
 
 const uidKey = 'sourcegraphAnonymousUid'
