+++ web/src/tracking/withActivation.tsx
@@ -11,7 +11,8 @@ import {
 import { dataOrThrowErrors, gql } from '../../../shared/src/graphql/graphql'
 import * as GQL from '../../../shared/src/graphql/schema'
 import { queryGraphQL } from '../backend/graphql'
-import { logUserEvent, UserEvent } from '../user/settings/backend'
+import { logUserEvent } from '../user/settings/backend'
+import { UserEvent } from '../../../browser/src/shared/backend/userEvents'
 
 /**
  * Fetches activation status from server.
