+++ web/src/user/settings/backend.tsx
@@ -85,36 +85,13 @@ export function setUserEmailVerified(user: GQL.ID, email: string, verified: bool
     )
 }
 
-// A user event.
-export enum UserEvent {
-    PAGEVIEW = 'PAGEVIEW',
-    SEARCHQUERY = 'SEARCHQUERY',
-    CODEINTEL = 'CODEINTEL',
-    CODEINTELREFS = 'CODEINTELREFS',
-    CODEINTELINTEGRATION = 'CODEINTELINTEGRATION',
-    CODEINTELINTEGRATIONREFS = 'CODEINTELINTEGRATIONREFS',
-
-    // Product stages
-    STAGEMANAGE = 'STAGEMANAGE',
-    STAGEPLAN = 'STAGEPLAN',
-    STAGECODE = 'STAGECODE',
-    STAGEREVIEW = 'STAGEREVIEW',
-    STAGEVERIFY = 'STAGEVERIFY',
-    STAGEPACKAGE = 'STAGEPACKAGE',
-    STAGEDEPLOY = 'STAGEDEPLOY',
-    STAGECONFIGURE = 'STAGECONFIGURE',
-    STAGEMONITOR = 'STAGEMONITOR',
-    STAGESECURE = 'STAGESECURE',
-    STAGEAUTOMATE = 'STAGEAUTOMATE',
-}
-
 /**
  * Log a user action (used to allow site admins on a Sourcegraph instance
  * to see a count of unique users on a daily, weekly, and monthly basis).
  *
  * Not used at all for public/sourcegraph.com usage.
  */
-export function logUserEvent(event: UserEvent): void {
+export function logUserEvent(event: string): void {
     if (window.context && window.context.sourcegraphDotComMode) {
         return
     }
